Rename simultaneous_tuning.py to simultaneous.py
